Numerous improvements and unit tests
Added local block labels, symbol file, #break, #continue, #IfCondSR

+++ run.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
 
+rm $1.obj.gz
 java -cp target/CPUSim64V2-1.0-SNAPSHOT.jar Assembler $1.asm $2 $3 $4 $5 $6 $7 $8 $9
-java -cp target/CPUSim64V2-1.0-SNAPSHOT.jar Simulation $1.obj.gz --verbose
+if [[ -e "$1.obj.gz" ]] ; then
+	java -cp target/CPUSim64V2-1.0-SNAPSHOT.jar Simulation $1.obj.gz --verbose
+fi
